Quite a few changes. Tooltips have been added for the buttons on the profile and the overview. Also a measure of keyboard control has been added:
w,s,a,d move the view of the profile and the overview.
W,S,A,D move the profile and fence on the overview and the fence on the profile. SPECIAL: When using a slanted profile then the direction of the movement will be forward and backward and so on relative to the viewpoint of the profile.
R,V,Q,E and r,v,q,e when on the profile view, "scroll" the profile as W,S,A,D does on the overview with a profile.
F and f toggle the fencing mode on the profile and the overview.
X and x toggle the profiling mode on the overview.
P,p and <space> get a profile from the overview, on both the profile and overview.
C,c,K,k and <return> classify the points on the profile selected by a fence.
I,i,O and o zoom in and out on both the profile and the overview.
T and t toggle whether the the fences and profiles should be orthogonal or slanted with respect to the window, for both the profile and overview.
Z and z redraw both the profile and overview.


git-svn-id: http://rsg.pml.ac.uk/intranet/svnroots/arsf-internal/trunk/tools/lag@746 e4e130c3-e29c-4adf-81f0-605662ec9256