This commit marks the migration from svn to git.
Adds the lidarquadtree library as a subdirectory of lag, which will
remain this way for the time being.

There seems to be a tangle of differences between the latest svn commit
and the earlier git commit, but the changes should sort themselves out a
few commits down the line. Please be aware though that the history at
this point is imperfect.